Add quickfix for OMP->CUDA not respecting dependencies
With the basic hybrid scheduler, dependencies from OMP to CUDA were not respected.
Memory transfers would happen before the kernel it depended on had completed. Waiting for dependencies would happen
on a worker thread, and the main thread would go ahead and submit the memory transfer even though the wait
had not completed.